Use original precision when casting coordinate values
If a float value is casted toa  double, a widening conversion
happens. This means that the source values is preserved
but the representation might change. We don't want this as it
gives too many decimal digits for geometry coordinate information.

This commit changes the behaviour so that the native precision
is honoured. For example, if the first longitude value has precision
of 4 decimal digits, the same precision is kept after casting.

By default this is switched off, since it will likely require all
netcdf geometries in radon to be updated.

The functionality can be enabled by setting an env variable:

FMINC_USE_IMPROVED_PRECISION=1